// ui-react/apps/console/src/client/types.gen.ts
export type UserUsername = string;

/**
 * User's password.
 */
export type UserPassword = string;

/**
 * JWT Token
 */
export type Jwt = string;

/**
 * User's ID.
 */
export type UserId = string;

/**
 * Specifies the method the user employed to register with ShellHub.
 */
export type UserOrigin = 'local';

/**
 * User's name.
 */
export type UserName = string;

/**
 * User's E-mail.
 */
export type UserEmail = string;

/**
 * Namespace's tenant ID
 */
export type NamespaceTenantId = string;

/**
 * Namespace's member role
 */
export type NamespaceMemberRole = 'administrator' | 'operator' | 'observer' | 'owner';

/**
 * Indicates the maximum number of namespaces a user is allowed to create. If set to 0, the user is not permitted to create any namespaces. If set to -1, the user has no limit on the number of namespaces they can create.
 */
export type MaxNamespaces = number;

export type UserAuth = {
    token: Jwt;
    id: UserId;
    origin: UserOrigin;
    user: UserUsername;
    name: UserName;
    email: UserEmail;
    /**
     * The recovery email serves as the user's final recourse to regain access
     * to their account.
     *
     */
    recovery_email: string;
    tenant: NamespaceTenantId;
    role: NamespaceMemberRole;
    /**
     * Indicates whether the user has MFA enabled.
     */
    mfa: boolean;
    /**
     * Indicates whether the user has admin status.
     */
    admin: boolean;
    max_namespaces: MaxNamespaces;
};